[core] update: ImageBlock 엔티티 수정 - S3 업로드 된 링크 저장하도록 변경
